+++ packages/twenty-server/src/database/commands/upgrade-version-command/workspace-command-provider.module.ts
@@ -25,6 +25,7 @@ import { V2_21_UpgradeVersionCommandModule } from 'src/database/commands/upgrade
 import { V2_22_UpgradeVersionCommandModule } from 'src/database/commands/upgrade-version-command/2-22/2-22-upgrade-version-command.module';
 import { V2_23_UpgradeVersionCommandModule } from 'src/database/commands/upgrade-version-command/2-23/2-23-upgrade-version-command.module';
 import { V2_25_UpgradeVersionCommandModule } from 'src/database/commands/upgrade-version-command/2-25/2-25-upgrade-version-command.module';
+import { V2_26_UpgradeVersionCommandModule } from 'src/database/commands/upgrade-version-command/2-26/2-26-upgrade-version-command.module';
 
 @Module({
   imports: [
@@ -53,6 +54,7 @@ import { V2_25_UpgradeVersionCommandModule } from 'src/database/commands/upgrade
     V2_22_UpgradeVersionCommandModule,
     V2_23_UpgradeVersionCommandModule,
     V2_25_UpgradeVersionCommandModule,
+    V2_26_UpgradeVersionCommandModule,
   ],
 })
 export class WorkspaceCommandProviderModule {}
